refactor: dashboard modal 파일 폴더 분리 (#25)
